requirements.txt
python-telegram-bot==22.7
aiohttp==3.9.5
paramiko==3.4.0
python-dotenv==1.0.1
